refactor: split out wild16 entrypoint and tests
